refactor: simplify Toaster struct and related methods
